# 🚀 Comprehensive Performance Optimizations for 2025