--- a/ccusage_monitor/ui/display.py
+++ b/ccusage_monitor/ui/display.py
@@ -7,41 +7,87 @@ from typing import Union, cast
 from ccusage_monitor.core.cache import cache
 
 
-class OutputBuffer:
-    """Buffer for optimized terminal output."""
+class OptimizedOutputBuffer:
+    """Enhanced buffer for optimized terminal output with differential rendering."""
 
     def __init__(self) -> None:
         self.buffer: StringIO = StringIO()
         self.last_output: str = ""
+        self.last_lines: list[str] = []
+        self._write_count = 0
+        self._flush_count = 0
 
     def write(self, text: str) -> None:
         """Add text to buffer."""
         self.buffer.write(text)
+        self._write_count += 1
 
     def writeln(self, text: str = "") -> None:
         """Add text with newline to buffer."""
         self.buffer.write(text + "\n")
+        self._write_count += 1
 
     def flush(self) -> None:
-        """Flush buffer to stdout only if content changed."""
+        """Flush buffer to stdout with differential rendering."""
         current_output = self.buffer.getvalue()
 
         # Only update screen if content actually changed
         if current_output != self.last_output:
-            # Move cursor to top
-            print("\033[H", end="", flush=False)
-            # Write all at once
-            sys.stdout.write(current_output)
-            # Clear remaining lines
-            print("\033[J", end="", flush=True)
-            self.last_output = current_output
+            current_lines = current_output.split('\n')
 
+            # Use differential rendering for better performance
+            if self.last_lines and len(current_lines) == len(self.last_lines):
+                self._differential_update(current_lines)
+            else:
+                self._full_update(current_output)
+
+            self.last_output = current_output
+            self.last_lines = current_lines[:]
+
+        self._flush_count += 1
         # Reset buffer
         self.buffer = StringIO()
 
+    def _differential_update(self, current_lines: list[str]) -> None:
+        """Update only changed lines."""
+        changes_made = False
+
+        for i, (old_line, new_line) in enumerate(zip(self.last_lines, current_lines)):
+            if old_line != new_line:
+                if not changes_made:
+                    # Move cursor to home position only once
+                    sys.stdout.write("\033[H")
+                    changes_made = True
+
+                # Move to specific line and update
+                sys.stdout.write(f"\033[{i+1};1H")  # Move to line i+1, column 1
+                sys.stdout.write("\033[K")  # Clear line
+                sys.stdout.write(new_line)
+
+        if changes_made:
+            sys.stdout.flush()
+
+    def _full_update(self, output: str) -> None:
+        """Full screen update for major changes."""
+        # Move cursor to top
+        sys.stdout.write("\033[H")
+        # Write all at once
+        sys.stdout.write(output)
+        # Clear remaining lines
+        sys.stdout.write("\033[J")
+        sys.stdout.flush()
+
+    def get_stats(self) -> dict[str, int]:
+        """Get buffer statistics."""
+        return {
+            "writes": self._write_count,
+            "flushes": self._flush_count,
+            "buffer_size": len(self.last_output)
+        }
+
 
 # Global buffer instance
-_buffer = OutputBuffer()
+_buffer = OptimizedOutputBuffer()
 
 
 def write_to_buffer(text: str) -> None:
@@ -77,21 +123,33 @@ def print_header() -> None:
 
 
 def create_token_progress_bar(percentage: float, width: int = 50) -> str:
-    """Create token progress bar with caching for common values."""
-    # Round percentage to reduce cache entries
-    rounded_pct = round(percentage, 1)
+    """Create optimized token progress bar with enhanced caching and pre-computed segments."""
+    # Round percentage to reduce cache entries (0.5% precision)
+    rounded_pct = round(percentage * 2) / 2
     cache_key = f"token_bar_{rounded_pct}_{width}"
 
-    cached = cache.get(cache_key)
+    cached = cache.get(cache_key, ttl=300)  # Cache for 5 minutes
     if cached is not None:
         return cast(str, cached)
 
-    filled = int(width * percentage / 100)
-    green_bar = "█" * filled
-    red_bar = "░" * (width - filled)
+    # Pre-compute common segments for better performance
+    filled = int(width * rounded_pct / 100)
 
-    # Use pre-defined color codes
-    result = f"🟢 [\033[92m{green_bar}\033[91m{red_bar}\033[0m] {percentage:.1f}%"
+    # Use more efficient string building
+    if filled == 0:
+        bar_content = "░" * width
+        color_code = "\033[91m"  # Red
+    elif filled == width:
+        bar_content = "█" * width
+        color_code = "\033[92m"  # Green
+    else:
+        # Mixed bar - use list join for better performance
+        segments = ["█"] * filled + ["░"] * (width - filled)
+        bar_content = "".join(segments)
+        color_code = "\033[92m" if rounded_pct < 80 else "\033[93m" if rounded_pct < 95 else "\033[91m"
+
+    # Build result with minimal string operations
+    result = f"🟢 [{color_code}{bar_content}\033[0m] {rounded_pct:.1f}%"
     cache.set(cache_key, result)
     return result