… PERFORMANCE_OPTIMIZATIONS.md
… performance_monitor.py
… test_optimizations.py